@@ -19,7 +19,7 @@
 		"@skeletonlabs/skeleton": "3.0.0-next.10",
 		"@skeletonlabs/skeleton-svelte": "1.0.0-next.17",
 		"@sveltejs/adapter-cloudflare": "^5.0.1",
-		"@sveltejs/kit": "^2.15.3",
+		"@sveltejs/kit": "^2.16.0",
 		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"autoprefixer": "^10.4.20",
 		"embla-carousel-svelte": "^8.5.2",
@@ -30,7 +30,7 @@
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.3",
 		"prettier-plugin-tailwindcss": "^0.6.10",
-		"svelte": "^5.18.0",
+		"svelte": "^5.19.0",
 		"svelte-check": "^4.1.4",
 		"svelte-flags": "^1.0.3",
 		"tailwindcss": "^3.4.17",
